sql: ban of REINDEX syntax
This change cut REINDEX syntax from keywords and also disables tests
which were using this syntax.

This patch affects only parser, REINDEX implementation still exists,
because it is planned to be fixed and enabled in the scope of #3195.

Note that REINDEX keyword is still in '%fallback' directive in
'parse.y'. The reason described in 'parse.y' in comment related
to '%fallback'.

Fixes #2174